+++ src/main/java/kr/codesqaud/cafe/controller/UserController.java
@@ -1,6 +1,6 @@
 package kr.codesqaud.cafe.controller;
 
-import kr.codesqaud.cafe.domain.UserLoginDTO;
+import kr.codesqaud.cafe.dto.UserLoginDTO;
 import kr.codesqaud.cafe.repository.member.MemberRepository;
 import kr.codesqaud.cafe.domain.User;
 import kr.codesqaud.cafe.validation.UserJoinValidator;
